fixed condition #2

+++ src/main/java/org/aueb/ds/pubsub/Publisher.java
@@ -98,7 +98,7 @@ public class Publisher extends AppNode implements Runnable, Serializable {
         File cwd = new File(System.getProperty("user.dir"));
         for (File video : cwd.listFiles()) {
             if (video.getName().contains(".mp4")
-                    || video.getName().replace(".mp4", "").split("#")[0].equals(actualName)) {
+                    && video.getName().replace(".mp4", "").split("#")[0].equals(actualName)) {
                 video.renameTo(new File(video.getName().replace(".mp4", ".removed")));
                 return;
             }
